:recycle: :: groupUser 일급컬렉션 제거
